Improve call to Analysis.setup
---
Merge from main to update improvements to Parts.py, MultiParts.py and Actors.py
---
Allow for __setitem__ callback in setup.props so the solution
setup in the app is updated when props are set. This avoids the need for an extra call
to update().
---
Changed "solutiontype" to "setuptype" in modules.SolveSetup.py to improve consistency.
---
Ongoing updtes to the HFSS_77GHz_Slot_Antenna.py example (not done yet)